adds first route
