+++ src/routes/item-prices.ts
@@ -1,0 +1,13 @@
+import Fastify, { FastifyReply, FastifyRequest } from 'fastify'
+import { db } from "../lib/db.js"
+
+export async function getLatestPrices(request: FastifyRequest, reply: FastifyReply): Promise<any> {
+  let response: any[] = []
+  const items = db.data?.items ?? []
+  for (let item of items) {
+    const { name, recordedPrices } = item;
+    let data = recordedPrices[recordedPrices.length - 1]
+    response.push({name, data})
+  }
+  reply.send(response)
+}
